move io interface stuff to real
