+++ src/main/java/frc/robot/elevator/ElevatorIOReal.java
@@ -14,8 +14,27 @@ import com.ctre.phoenix6.signals.NeutralModeValue;
 import edu.wpi.first.units.measure.Current;
 import edu.wpi.first.units.measure.Temperature;
 import edu.wpi.first.units.measure.Voltage;
+import org.littletonrobotics.junction.AutoLog;
 
-public class ElevatorIOReal implements ElevatorIO {
+public class ElevatorIOReal {
+
+  @AutoLog
+  public static class ElevatorIOInputs {
+    public double leaderPositionMeters = 0.0;
+    public double leaderVelocityMetersPerSec = 0.0;
+    public double leaderStatorCurrentAmps = 0.0;
+    public double leaderSupplyCurrentAmps = 0.0;
+    public double leaderVoltage = 0.0;
+    public double leaderTempC = 0.0;
+
+    public double followerPositionMeters = 0.0;
+    public double followerVelocityMetersPerSec = 0.0;
+    public double followerStatorCurrentAmps = 0.0;
+    public double followerSupplyCurrentAmps = 0.0;
+    public double followerVoltage = 0.0;
+    public double followerTempC = 0.0;
+  }
+
   protected TalonFX leader = new TalonFX(10, "*");
   protected TalonFX follower = new TalonFX(11, "*");
 
@@ -101,7 +120,6 @@ public class ElevatorIOReal implements ElevatorIO {
     follower.optimizeBusUtilization();
   }
 
-  @Override
   public void updateInputs(ElevatorIOInputs inputs) {
     BaseStatusSignal.refreshAll(
         leaderPositionMeters,
@@ -132,26 +150,26 @@ public class ElevatorIOReal implements ElevatorIO {
     inputs.followerTempC = followerTemp.getValueAsDouble();
   }
 
-  @Override
   public void setVoltage(double volts) {
     leader.setControl(voltageOut.withOutput(volts));
   }
 
-  @Override
   public void setCurrent(double amps) {
     leader.setControl(torqueCurrent.withOutput(amps));
   }
 
-  @Override
   public void setPositionSetpoint(double positionMeters, double acceleration) {
     positionSetpoint = positionMeters;
     leader.setControl(
         motionMagicVoltage.withPosition(positionMeters).withAcceleration(acceleration));
   }
 
-  @Override
   public void resetEncoder(double position) {
     leader.setPosition(position);
     follower.setPosition(position);
+  }
+
+  public void stop() {
+    setVoltage(0.0);
   }
 }
